Changed health to Slider -Added a health slider to the GUICanvas in the room.  I kept it simple, we can 100% change the color, give it a background color, change the font, etc.  I wanted to check in this version before trying to do something more elaborate.  Feedback is welcomed. -removed PlayerHealthHUD from the debug menu.  The prefab still exists in Player folder if we want to go back to it (though I don’t know if people have been making changes to it, there's also a prefab in tots sandbox from when he was debugging things earlier.) -added the gameover song to the game over screen because I noticed it was throwing an exception since no song was assigned.
